refactor(checker): convert dispatch to directory module, move companions inside
The check-crate-root-files policy disallows new root-level modules in
crates that already exceed 4 non-facade root files. tsz-checker had 8
before this change.

Convert dispatch.rs → dispatch/mod.rs and move all companion files into
the same directory:

  dispatch_helpers.rs → dispatch/helpers.rs
  dispatch_jsx.rs     → dispatch/jsx.rs
  dispatch_this.rs    → dispatch/this.rs
  dispatch_yield.rs   → dispatch/yield_.rs

Update imports in each companion (super::dispatch::ExpressionDispatcher
→ super::ExpressionDispatcher), add the four sub-module declarations to
dispatch/mod.rs, remove the now-redundant declarations from lib.rs, and
fix the two cross-crate references to dispatch_helpers::keyword_type_mapping.

Zero behavior change. Verified: cargo check -p tsz-checker passes.
