+++ crates/tsz-checker/src/lib.rs
@@ -19,10 +19,6 @@ extern crate self as tsz_checker;
 
 pub mod context;
 pub mod dispatch;
-mod dispatch_helpers;
-mod dispatch_jsx;
-mod dispatch_this;
-mod dispatch_yield;
 pub mod error_reporter;
 pub mod expr;
 pub mod module_resolution;
